No use of 'friend' and made some getters/setters
Subject asks not to make use of the friend qualifier. Normally it makes
it easy to make functions be able to access private members. I've gotten
rid of the friend qualifier previously used for the '<<' overload. To
make it work with privates, you just have to define it out-of-line
(outside the class decl) and make sure to only use public accessable
members.

I've gone and made all Contact members private, making getters and
setters for each of them. It made the 'add_contact' a bit less readable
because you can't pass the field to a function anymore (it's private).