Finished ex02
This was pretty easy after figuring out what was asked of you, how the
log file tied in to the tests.cpp file. We were tasked to create a
Account.cpp file matching the Account.hpp file to produce the right
output of logs. It meant just printing the correct values in the correct
format. Just increment and set the right values.

What was pretty interesting to learn is how static members and
properties work in classes. They are meant to within the scope of the
class, still alive when code is not currently there. You can almost see
it as having variables or functions defined inside a namespace (that
being the name of the class). It can be really handy to be able to call
a static member function without having to make an object for it. This
is seen in the tests.cpp file calling Account::displayAccountInfos().